refactor: move sidebar port helpers to background/helpers/sidebarPort
Extract setSidebarPort, clearSidebarPort, getSidebarPort from
freighterApiMessageListener into their own module to reduce coupling.
Update all imports across source and test files.
