+++ extension/src/background/messageListener/popupMessageListener.ts
@@ -29,7 +29,7 @@ import {
   activeQueueUuids,
   sidebarQueueUuids,
 } from "background/helpers/queueCleanup";
-import { getSidebarPort } from "./freighterApiMessageListener";
+import { getSidebarPort } from "background/helpers/sidebarPort";
 
 import { fundAccount } from "./handlers/fundAccount";
 import { createAccount } from "./handlers/createAccount";
